test(review): locate hot intake assertions with scheduler
